requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
filelock==3.8.0
huggingface-hub==0.10.1
idna==3.4
numpy==1.23.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
timm==0.6.11
tokenizers==0.13.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.23.1
typing-extensions==4.4.0
urllib3==1.26.12
